Minor changes in code structure

+++ app/implementation.py
@@ -1,20 +1,24 @@
 from passlib.hash import sha256_crypt
 from dbquery import querydb
+from schema import Schema, And, Use
 
 
 def getAllUsers():
-
     data = 'SELECT * FROM "user"'
     return querydb(data, operation='GET', check='list')
 
 
 def postUser(request):
+    passwd = sha256_crypt.encrypt(str(request.json['password']))
+    data = "INSERT INTO \"user\" (first_name, last_name, username, email, admin, password) values ('" + str(
+        request.json['first_name']) + "', '" + str(request.json['last_name']) + "', '" + str(
+        request.json['username']) + "', '" + str(request.json['email']) + "', '" + str(
+        request.json['admin']) + "', '" + str(passwd) + "')"
 
-    return querydb("", operation='POST', request=request)
+    return querydb(data, operation='POST')
 
 
 def getUserByUsername(username):
-
     data = 'SELECT * FROM "user" where username = ' + "'" + str(username) + "'"
     data = querydb(data, operation='GET', check='tuple', )
     if data == "No data to return.":
@@ -23,20 +27,37 @@ def getUserByUsername(username):
 
 
 def getUserByID(user_id):
-
     data = 'SELECT * FROM "user" where user_id = ' + str(user_id)
     return querydb(data, 'GET', 'tuple', user_id=user_id)
 
 
 def putUserByID(request, user_id):
 
-    return querydb("", 'PUT', user_id=user_id, request=request)
+    data = getUserByID(user_id)
+    if data == "No data to return.":
+        return postUser(request), 204
+    else:
+
+        putData = putDataCheck(request, data)
+        if putData == "Something went wrong in mapping data.":
+            return {"msg": "Something went wrong in mapping data."}, 500
+
+        data = "UPDATE \"user\" SET user_id = '" + str(user_id) + "', first_name = '" + putData[
+            0] + "', last_name = '" + \
+               putData[1] + "', username = '" + putData[2] + "', email = '" + putData[3] + "', admin = '" + putData[4] + \
+               "', password = '" + putData[5] + "' WHERE user_id = '" + str(user_id) + "'"
+
+        return querydb(data, 'PUT', user_id=user_id)
 
 
 def deleteUserByID(user_id):
 
-    data = 'DELETE FROM "user" WHERE user_id = ' + (str(user_id))
-    return querydb(data, 'DELETE', user_id=user_id)
+    data = getUserByID(user_id)
+    if data == "No data to return.":
+        return {"msg": "User with user_id " + str(user_id) + " does not exist in DB."}
+    else:
+        data = 'DELETE FROM "user" WHERE user_id = ' + (str(user_id))
+        return querydb(data, 'DELETE', user_id=user_id)
 
 
 def putDataCheck(request, data):
@@ -71,3 +92,11 @@ def putDataCheck(request, data):
         print(e)
         return "Something went wrong in mapping data."
 
+
+def getSchema():
+    return Schema({'first_name': And(str, len),
+                     'last_name': And(str, len),
+                     'username': And(str, len),
+                     'email': And(str, len),
+                     'admin': And(Use(bool)),
+                     'password': And(str, len)})
